move version check functions to a separate class to reduce code duplication
